Implemented suggestions as per PR #18
