+++ src/CarbonIntensityTime.DataFactory/FactoryClient.cs
@@ -7,6 +7,9 @@ using Microsoft.Azure.Management.DataFactory.Models;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.Rest;
+using Microsoft.VisualBasic;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 using PipelineRun = CarbonIntensityTime.DataFactory.Models.PipelineRun;
 
 namespace CarbonIntensityTime.DataFactory;
@@ -83,7 +86,7 @@ public class FactoryClient : IFactoryClient
                     Start = ar.ActivityRunStart,
                     End = ar.ActivityRunEnd,
                     Duration = ar.DurationInMs,
-                    Input = ar.Input,
+                    Input = PipelineActivity.ConvertToDictionary(ar.Input),
                     PipelineName = ar.PipelineName,
                     PipelineRunId = ar.PipelineRunId
                 }).ToList()
